[수정] Day 1: 부동 소수점 오류 위험 수정
- double -> BigDecimal 타입 사용

+++ src/main/java/example/dailycoding/coupon/dto/CouponDto.java
@@ -22,7 +22,7 @@ public class CouponDto {
                 coupon.getId(),
                 coupon.getName(),
                 coupon.getDiscountType(),
-                coupon.getDiscount(),
+                coupon.getDiscount().longValue(),
                 coupon.getStartDate(),
                 coupon.getEndDate()
         );
